--- a/MANIFEST.in
+++ b/MANIFEST.in
@@ -1,4 +1,3 @@
-include .flake8
 include .gitignore
 include CREDITS
 include HISTORY.rst