--- a/demo/unix_daemon.py
+++ b/demo/unix_daemon.py
@@ -5,7 +5,7 @@
 # found in the LICENSE file.
 
 """A basic unix daemon using the python-daemon library:
-http://pypi.python.org/pypi/python-daemon
+http://pypi.python.org/pypi/python-daemon.
 
 Example usages: