--- a/pyftpdlib/_compat.py
+++ b/pyftpdlib/_compat.py
@@ -9,7 +9,7 @@ import sys
 import types
 
 
-PY3 = sys.version_info[0] == 3
+PY3 = sys.version_info[0] >= 3
 _SENTINEL = object()
 
 if PY3: