--- a/pyftpdlib/log.py
+++ b/pyftpdlib/log.py
@@ -123,9 +123,8 @@ class LogFormatter(logging.Formatter):
             message = repr(record.message)
 
         formatted = prefix + " " + message
-        if record.exc_info:
-            if not record.exc_text:
-                record.exc_text = self.formatException(record.exc_info)
+        if record.exc_info and not record.exc_text:
+            record.exc_text = self.formatException(record.exc_info)
         if record.exc_text:
             formatted = formatted.rstrip() + "\n" + record.exc_text
         return formatted.replace("\n", "\n    ")