--- a/scripts/internal/git_pre_commit.py
+++ b/scripts/internal/git_pre_commit.py
@@ -28,7 +28,7 @@ import sys
 
 
 PYTHON = sys.executable
-PY3 = sys.version_info[0] == 3
+PY3 = sys.version_info[0] >= 3
 THIS_SCRIPT = os.path.realpath(__file__)