--- a/scripts/internal/winmake.py
+++ b/scripts/internal/winmake.py
@@ -186,7 +186,7 @@ def safe_rmtree(path):
 
 def recursive_rm(*patterns):
     """Recursively remove a file or matching a list of patterns."""
-    for root, subdirs, subfiles in os.walk(u'.'):
+    for root, subdirs, subfiles in os.walk('.'):
         root = os.path.normpath(root)
         if root.startswith('.git/'):
             continue