See: https://blog.jerrycodes.com/ruff-the-python-linter/. Advantages:

1) an order of magnitude faster than flake8 + isort (0.1 secs instead of 2.0 secs):

Before:
```
$ time make flake8 isort
real	0m2,042s
user	0m7,568s
sys	0m0,380s
```

Now:
```
$ time make ruff 
real	0m0,127s
user	0m0,183s
sys	0m0,045s
```

2) Many more code quality checks than before.